requirements/test.txt
-r prod.txt
pylint==1.9.2
pytest==3.5.1
pytest-flask==0.10.0
